+++ qa/tasks/mgr/dashboard/test_erasure_code_profile.py
@@ -99,7 +99,7 @@ class ECPTest(DashboardTestCase):
         self.assertStatus(204)
 
     def test_ecp_info(self):
-        self._get('/api/erasure_code_profile/_info')
+        self._get('/ui-api/erasure_code_profile/info')
         self.assertSchemaBody(JObj({
             'names': JList(six.string_types),
             'failure_domains': JList(six.string_types),
