+++ src/pybind/mgr/dashboard/controllers/erasure_code_profile.py
@@ -3,22 +3,10 @@ from __future__ import absolute_import
 
 from cherrypy import NotFound
 
-from . import ApiController, RESTController, Endpoint, ReadPermission
+from . import ApiController, RESTController, Endpoint, ReadPermission, UiApiController
 from ..security import Scope
 from ..services.ceph_service import CephService
 from .. import mgr
-
-
-def _serialize_ecp(name, ecp):
-    def serialize_numbers(key):
-        value = ecp.get(key)
-        if value is not None:
-            ecp[key] = int(value)
-
-    ecp['name'] = name
-    serialize_numbers('k')
-    serialize_numbers('m')
-    return ecp
 
 
 @ApiController('/erasure_code_profile', Scope.POOL)
@@ -29,17 +17,14 @@ class ErasureCodeProfile(RESTController):
     '''
 
     def list(self):
-        ret = []
-        for name, ecp in mgr.get('osd_map').get('erasure_code_profiles', {}).items():
-            ret.append(_serialize_ecp(name, ecp))
-        return ret
+        return CephService.get_erasure_code_profiles()
 
     def get(self, name):
-        try:
-            ecp = mgr.get('osd_map')['erasure_code_profiles'][name]
-            return _serialize_ecp(name, ecp)
-        except KeyError:
-            raise NotFound('No such erasure code profile')
+        profiles = CephService.get_erasure_code_profiles()
+        for p in profiles:
+            if p['name'] == name:
+                return p
+        raise NotFound('No such erasure code profile')
 
     def create(self, name, **kwargs):
         profile = ['{}={}'.format(key, value) for key, value in kwargs.items()]
@@ -49,9 +34,12 @@ class ErasureCodeProfile(RESTController):
     def delete(self, name):
         CephService.send_command('mon', 'osd erasure-code-profile rm', name=name)
 
+
+@UiApiController('/erasure_code_profile', Scope.POOL)
+class ErasureCodeProfileUi(ErasureCodeProfile):
     @Endpoint()
     @ReadPermission
-    def _info(self):
+    def info(self):
         '''Used for profile creation and editing'''
         config = mgr.get('config')
         osd_map_crush = mgr.get('osd_map_crush')
